add a way to tell actions they've been registered
Within applications actions may need to refer to other actions and stores. The
app factory may dynamically create such instances, which prevents an action from
having design time references.

Actions may optionally implement a register() method, which is passed a registry
object. Within an application it should be possible to cast this object to a
known type, and then gain references to actions and stores.

A handle may be returned to allow for cleanup when the action is deregistered.